+++ spec/factories/colleges.rb
@@ -6,5 +6,9 @@ FactoryBot.define do
     admin_email { "admin@#{name}.edu" }
     subdomain { name.downcase }
     dean { "Dean #{name} " }
+    allow_clipping true
+    restrict_clipping_group_size false
+    advantage_clips false
+    size_sort { 'no_sort' }
   end
 end