+++ spec/factories/draws.rb
@@ -3,8 +3,6 @@
 FactoryBot.define do
   factory :draw do
     name 'MyString'
-    allow_clipping true
-    restrict_clipping_group_size false
     active true
 
     factory :draw_with_members do